pointcloud_opt_simp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cbor2==5.7.1
certifi==2025.10.5
click==8.3.0
fastapi==0.120.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
grpclib==0.4.8
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modal==1.2.1
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
rich==14.2.0
setuptools==80.9.0
shellingham==1.5.4
sigtools==4.0.1
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
synchronicity==0.10.2
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
triton==3.5.0
typer==0.20.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
watchfiles==1.1.1
yarl==1.22.0
